Break hero into it's own component to be used elsewhere
